Split off range management code from PlotCanvas
