+++ bokehjs/src/lib/models/tools/actions/zoom_base_tool.ts
@@ -17,7 +17,7 @@ export class ZoomBaseToolView extends ActionToolView {
     const zoom_info = scale_range(frame, this.model.sign * this.model.factor, h_axis, v_axis)
 
     this.plot_view.push_state('zoom_out', {range: zoom_info})
-    this.plot_view.update_range(zoom_info, false, true)
+    this.plot_view.update_range(zoom_info, {scrolling: true})
 
     if (this.model.document)
       this.model.document.interactive_start(this.plot_model)
