+++ bokehjs/src/lib/models/tools/gestures/wheel_zoom_tool.ts
@@ -45,7 +45,9 @@ export class WheelZoomToolView extends GestureToolView {
     const zoom_info = scale_range(frame, factor, h_axis, v_axis, {x: sx, y: sy})
 
     this.plot_view.push_state('wheel_zoom', {range: zoom_info})
-    this.plot_view.update_range(zoom_info, false, true, this.model.maintain_focus)
+
+    const {maintain_focus} = this.model
+    this.plot_view.update_range(zoom_info, {scrolling: true, maintain_focus})
 
     if (this.model.document != null)
       this.model.document.interactive_start(this.plot_model)
